server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["chasingamy"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pydantic-settings = "^2.9.1"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
tortoise-orm = "^0.25.1"
aerich = "^0.9.0"
asyncpg = "^0.30.0"
tomlkit = "^0.13.3"
redis = "^6.2.0"

[tool.poetry.dev-dependencies]

[tool.aerich]
tortoise_orm = "config_reader.TORTOISE_ORM"
location = "./db/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
